Refactor extract class
