+++ FBReader/src/main/java/org/geometerplus/android/fbreader/benetech/AbstractReadingListTabContainer.java
@@ -1,0 +1,7 @@
+package org.geometerplus.android.fbreader.benetech;
+
+/**
+ * Created by animal@martus.org on 4/27/16.
+ */
+public class AbstractReadingListTabContainer {
+}
